Large set of changes to support a Mapping Africa qualification test. This first pass includes the creation of the qualification type and its inclusion into the creation of HITs as a requirement. It also includes new code to support a training module, which loosely connects to MTurk.
Still to do is to handle the case where the score is a failing score; also the call to the scoring routine needs to be added to this code. Plus we still need a daemon that polls for MTurk Qualification Requests, which result from the user taking a qualification test. When one of these is found, we check that the user's training ID matches that of a user that successfully completed the training module, and then we grant the user the Mapping Africa qualification, which sends him an email notification, and allows him to start completing our HITs.