New utility to dispose of a single HIT by HIT ID.
Makes certain validity tests to try to ensure that HIT can safely be deleted,
but because, in some cases, Mturk failures caused the code to abort,
the utility can't always detect all the conditions where a deletion
should not be made without manual examination of the actual DB records 
and Mturk lists involved.

Consequently, the utility takes an optional '-l' (lower-case L) switch to run 
it in list-only mode. No actions will be taken in this mode.  It is recommended 
to use this switch first, and then follow-up as appropriate.